fix: specify Effect on `execute`, not rule collection
Because rules need to be deleted in the reverse order of creation,
`execute` needs the knowledge and the expansion of the action also
needs to be carried out by `execute`.

"Nested transactions" (ability to mix batches of deletions and
creations in a single commit) left as an exercise for the future.